requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
altair==4.2.0
anyio==3.6.2
appdirs==1.4.4
async-timeout==4.0.2
asyncer==0.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
coloredlogs==15.0.1
commonmark==0.9.1
cv2
decorator==5.1.1
entrypoints==0.4
fastapi==0.87.0
filelock==3.8.0
filetype==1.2.0
flatbuffers==22.10.26
frozenlist==1.3.3
gdown==4.5.1
gitdb==4.0.9
GitPython==3.1.29
h11==0.14.0
humanfriendly==10.0
idna==3.4
ImageHash==4.3.1
imageio==2.22.4
importlib-metadata==5.0.0
Jinja2==3.1.2
jsonschema==4.17.0
llvmlite==0.39.1
MarkupSafe==2.1.1
mpmath==1.2.1
multidict==6.0.2
networkx==2.8.8
numba==0.56.4
numpy==1.23.5
onnxruntime==1.13.1
opencv-python-headless==4.6.0.66
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
pooch==1.6.0
protobuf==3.20.3
pyarrow==10.0.0
pydantic==1.10.2
pydeck==0.8.0
Pygments==2.13.0
PyMatting==1.1.8
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
rembg==2.0.29
requests==2.28.1
rich==12.6.0
scikit-image==0.19.3
scipy==1.9.3
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
starlette==0.21.0
streamlit==1.15.1
sympy==1.11.1
tifffile==2022.10.10
toml==0.10.2
toolz==0.12.0
torch
torchaudio
torchvision
tornado==6.2
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
uvicorn==0.20.0
validators==0.20.0
watchdog==2.1.9
yarl==1.8.1
zipp==3.10.0